tests/test_operations.py
```python
import unittest
from unittest.mock import patch
from src import operations

class TestOperations(unittest.TestCase):

    @patch('src.operations.User')
    def test_create_user(self, mock_user):
        mock_user.return_value = {'user_id': '123', 'name': 'John Doe', 'investment_preferences': 'Commercial', 'risk_tolerance': 'High'}
        result = operations.create_user('123', 'John Doe', 'Commercial', 'High')
        self.assertEqual(result, {'user_id': '123', 'name': 'John Doe', 'investment_preferences': 'Commercial', 'risk_tolerance': 'High'})

    @patch('src.operations.Investment')
    def test_create_investment(self, mock_investment):
        mock_investment.return_value = {'investment_id': '456', 'property_id': '789', 'reit_id': '1011', 'investment_amount': '50000', 'investment_date': '2022-01-01'}
        result = operations.create_investment('456', '789', '1011', '50000', '2022-01-01')
        self.assertEqual(result, {'investment_id': '456', 'property_id': '789', 'reit_id': '1011', 'investment_amount': '50000', 'investment_date': '2022-01-01'})

    @patch('src.operations.Property')
    def test_create_property(self, mock_property):
        mock_property.return_value = {'property_id': '789', 'property_type': 'Commercial', 'property_value': '1000000', 'rental_income': '5000'}
        result = operations.create_property('789', 'Commercial', '1000000', '5000')
        self.assertEqual(result, {'property_id': '789', 'property_type': 'Commercial', 'property_value': '1000000', 'rental_income': '5000'})

    @patch('src.operations.REIT')
    def test_create_reit(self, mock_reit):
        mock_reit.return_value = {'reit_id': '1011', 'reit_name': 'Top REIT', 'reit_value': '2000000'}
        result = operations.create_reit('1011', 'Top REIT', '2000000')
        self.assertEqual(result, {'reit_id': '1011', 'reit_name': 'Top REIT', 'reit_value': '2000000'})

    @patch('src.operations.AI_Tools')
    def test_generate_recommendations(self, mock_ai_tools):
        mock_ai_tools.generate_recommendations.return_value = ['Investment 1', 'Investment 2', 'Investment 3']
        result = operations.generate_recommendations('123')
        self.assertEqual(result, ['Investment 1', 'Investment 2', 'Investment 3'])

    @patch('src.operations.AI_Tools')
    def test_automate_management(self, mock_ai_tools):
        mock_ai_tools.automate_management.return_value = True
        result = operations.automate_management('456')
        self.assertEqual(result, True)

    @patch('src.operations.AI_Tools')
    def test_analyze_data(self, mock_ai_tools):
        mock_ai_tools.analyze_data.return_value = {'user_behavior': 'Positive', 'areas_for_improvement': 'None'}
        result = operations.analyze_data('123')
        self.assertEqual(result, {'user_behavior': 'Positive', 'areas_for_improvement': 'None'})

if __name__ == '__main__':
    unittest.main()
```